Refactor parts of Service and Db into ServiceDb
This is part of an effort to encapsulate the database
specific implementation away from imdb::Service.
